+++ src-tauri/src/graphql/mod.rs
@@ -1,1 +1,2 @@
+pub mod datasource;
 pub mod generated;
