+++ src-tauri/src/main.rs
@@ -2,14 +2,15 @@
 // Prevents additional console window on Windows in release, DO NOT REMOVE!!
 #![cfg_attr(not(debug_assertions), windows_subsystem = "windows")]
 
-use crate::datasource::DataSource;
+use database::DirectoryManager;
+use graphql::datasource::DataSource;
 
-mod datasource;
+mod database;
 mod graphql;
 mod tauri_commands;
 
 fn main() {
-  DataSource::init_games();
+  DirectoryManager::init_games();
 
   let schema = async_graphql::Schema::build(
     graphql::generated::Query,
